Move Env to a separate module
