+++ integration/test/Testlib/App.hs
@@ -2,15 +2,6 @@
 {-# LANGUAGE RecordWildCards #-}
 {-# OPTIONS_GHC -Wno-unused-matches #-}
 
--- |
---
--- Table of Contents:
--- - SECTION_APP : App, Env, ServiceMap, Context
--- - SECTION_ASSERTIONS
--- - SECTION_JSON
--- - SECTION_REQUEST : requests and responses
--- - SECTION_PRINTING : Terminal and printingmodule App where
--- - SECTION_MODSERVICE : Start services with modified configuration
 module Testlib.App where
 
 import Control.Concurrent.Async (mapConcurrently_)
@@ -34,14 +25,12 @@ import qualified Data.ByteString.Lazy as L
 import qualified Data.ByteString.Lazy.Char8 as LC8
 import qualified Data.CaseInsensitive as CI
 import Data.Char
-import qualified Data.Char as C
 import Data.Foldable
 import Data.Function
 import Data.Functor
 import Data.IORef
 import Data.List
 import Data.List.Split (splitOn)
-import Data.Map (Map)
 import qualified Data.Map.Strict as Map
 import qualified Data.Scientific as Sci
 import Data.String
@@ -49,10 +38,8 @@ import Data.String.Conversions (cs)
 import qualified Data.Text as T
 import qualified Data.Text.Encoding as T
 import Data.Traversable
-import Data.Word
 import qualified Data.Yaml as Yaml
 import GHC.Exception
-import GHC.Generics
 import GHC.Records
 import GHC.Stack
 import qualified GHC.Stack as Stack
@@ -61,16 +48,11 @@ import qualified Network.HTTP.Types as HTTP
 import qualified Network.Socket as N
 import Network.URI (uriToString)
 import System.Directory
-import System.Exit (exitFailure)
-import System.FilePath (joinPath, takeDirectory, (</>))
-import System.FilePath.Posix (splitPath)
+import System.FilePath
 import System.IO
 import qualified System.IO.Error as Error
 import System.Process (CreateProcess (..), createProcess, proc, terminateProcess)
-
--------------------------------------------------------------------------------
--- - SECTION_APP : App, Env, ServiceMap, Context
--------------------------------------------------------------------------------
+import Testlib.Env
 
 newtype App a = App {unApp :: ReaderT Env IO a}
   deriving
@@ -136,201 +118,6 @@ runTest ge action = do
                         pure (Just (colored yellow (displayException e)))
                     )
                 ]
-
--- | Initialised once per test.
-data Env = Env
-  { serviceMap :: Map String ServiceMap,
-    domain1 :: String,
-    domain2 :: String,
-    defaultAPIVersion :: Int,
-    manager :: HTTP.Manager,
-    serviceConfigsDir :: FilePath,
-    servicesCwdBase :: Maybe FilePath,
-    prekeys :: IORef [(Int, String)],
-    lastPrekeys :: IORef [String]
-  }
-
--- | Initialised once per testsuite.
-data GlobalEnv = GlobalEnv
-  { gServiceMap :: Map String ServiceMap,
-    gDomain1 :: String,
-    gDomain2 :: String,
-    gDefaultAPIVersion :: Int,
-    gManager :: HTTP.Manager,
-    gServiceConfigsDir :: FilePath,
-    gServicesCwdBase :: Maybe FilePath
-  }
-
-data IntegrationConfig = IntegrationConfig
-  { backendOne :: BackendConfig,
-    backendTwo :: BackendConfig
-  }
-
-instance FromJSON IntegrationConfig where
-  parseJSON v =
-    IntegrationConfig
-      <$> parseJSON v
-      <*> withObject "ServiceMap at backendTwo" (Aeson..: fromString "backendTwo") v
-
-data ServiceMap = ServiceMap
-  { brig :: HostPort,
-    cannon :: HostPort,
-    cargohold :: HostPort,
-    federatorInternal :: HostPort,
-    federatorExternal :: HostPort,
-    galley :: HostPort,
-    gundeck :: HostPort,
-    nginz :: HostPort,
-    spar :: HostPort
-  }
-  deriving (Show, Generic)
-
-instance FromJSON ServiceMap
-
-data BackendConfig = BackendConfig
-  { beServiceMap :: ServiceMap,
-    originDomain :: String
-  }
-  deriving (Show)
-
-instance FromJSON BackendConfig where
-  parseJSON v =
-    BackendConfig
-      <$> parseJSON v
-      <*> withObject "BackendConfig" (\ob -> ob .: fromString "originDomain") v
-
-data HostPort = HostPort
-  { host :: String,
-    port :: Word16
-  }
-  deriving (Show, Generic)
-
-instance FromJSON HostPort
-
-data Service = Brig | Galley | Cannon
-  deriving (Show, Eq, Ord)
-
-serviceName :: Service -> String
-serviceName srv = map C.toLower (show srv)
-
-serviceHostPort :: ServiceMap -> Service -> HostPort
-serviceHostPort m Brig = m.brig
-serviceHostPort m Galley = m.galley
-serviceHostPort m Cannon = m.cannon
-
-mkGlobalEnv :: FilePath -> IO GlobalEnv
-mkGlobalEnv cfgFile = do
-  eith <- Yaml.decodeFileEither cfgFile
-  intConfig <- case eith of
-    Left err -> do
-      hPutStrLn stderr $ "Could not parse " <> cfgFile <> ": " <> Yaml.prettyPrintParseException err
-      exitFailure
-    Right (intConfig :: IntegrationConfig) -> pure intConfig
-
-  let devEnvProjectRoot = case splitPath (takeDirectory cfgFile) of
-        [] -> Nothing
-        ps ->
-          if last ps == "services"
-            then Just (joinPath (init ps))
-            else Nothing
-
-  let configsDir =
-        case devEnvProjectRoot of
-          Just root -> root </> "./services/.integration/A/etc/wire/"
-          Nothing -> "/etc/wire"
-
-  manager <- HTTP.newManager HTTP.defaultManagerSettings
-  pure
-    GlobalEnv
-      { gServiceMap =
-          Map.fromList
-            [ (intConfig.backendOne.originDomain, intConfig.backendOne.beServiceMap),
-              (intConfig.backendTwo.originDomain, intConfig.backendTwo.beServiceMap)
-            ],
-        gDomain1 = intConfig.backendOne.originDomain,
-        gDomain2 = intConfig.backendTwo.originDomain,
-        gDefaultAPIVersion = 4,
-        gManager = manager,
-        gServiceConfigsDir = configsDir,
-        gServicesCwdBase = devEnvProjectRoot <&> (</> "services")
-      }
-
-mkEnv :: GlobalEnv -> IO Env
-mkEnv ge = do
-  pks <- newIORef (zip [1 ..] somePrekeys)
-  lpks <- newIORef someLastPrekeys
-  pure
-    Env
-      { serviceMap = gServiceMap ge,
-        domain1 = gDomain1 ge,
-        domain2 = gDomain2 ge,
-        defaultAPIVersion = gDefaultAPIVersion ge,
-        manager = gManager ge,
-        serviceConfigsDir = gServiceConfigsDir ge,
-        servicesCwdBase = gServicesCwdBase ge,
-        prekeys = pks,
-        lastPrekeys = lpks
-      }
-
-somePrekeys :: [String]
-somePrekeys =
-  [ "pQABAQECoQBYIOjl7hw0D8YRNqkkBQETCxyr7/ywE/2R5RWcUPM+GJACA6EAoQBYILLf1TIwSB62q69Ojs/X1tzJ+dYHNAw4QbW/7TC5vSZqBPY=",
-    "pQABAQICoQBYIGoXawUQWQ9ZW+MXhvuo9ALOBUjLff8S5VdAokN29C1OA6EAoQBYILLf1TIwSB62q69Ojs/X1tzJ+dYHNAw4QbW/7TC5vSZqBPY=",
-    "pQABAQMCoQBYIEjdt+YWd3lHmG8pamULLMubAMZw556IO8kW7s1MLFytA6EAoQBYILLf1TIwSB62q69Ojs/X1tzJ+dYHNAw4QbW/7TC5vSZqBPY=",
-    "pQABAQQCoQBYIPIaOA3Xqfk4Lh2/pU88Owd2eW5eplHpywr+Mx4QGyiMA6EAoQBYILLf1TIwSB62q69Ojs/X1tzJ+dYHNAw4QbW/7TC5vSZqBPY=",
-    "pQABAQUCoQBYIHnafNR4Gh3ID71lYzToewEVag4EKskDFq+gaeraOlSJA6EAoQBYILLf1TIwSB62q69Ojs/X1tzJ+dYHNAw4QbW/7TC5vSZqBPY=",
-    "pQABAQYCoQBYIFXUkVftE7kK22waAzhOjOmJVex3EBTU8RHZFx2o1Ed8A6EAoQBYILLf1TIwSB62q69Ojs/X1tzJ+dYHNAw4QbW/7TC5vSZqBPY=",
-    "pQABAQcCoQBYIDXdN8VlKb5lbgPmoDPLPyqNIEyShG4oT/DlW0peRRZUA6EAoQBYILLf1TIwSB62q69Ojs/X1tzJ+dYHNAw4QbW/7TC5vSZqBPY=",
-    "pQABAQgCoQBYIJH1ewvIVV3yGqQvdr/QM9HARzMgo5ksOTRyKEuN2aZzA6EAoQBYILLf1TIwSB62q69Ojs/X1tzJ+dYHNAw4QbW/7TC5vSZqBPY=",
-    "pQABAQkCoQBYIFcAnXdx0M1Q1hoDDfgMK9r+Zchn8YlVHHaQwQYhRk1dA6EAoQBYILLf1TIwSB62q69Ojs/X1tzJ+dYHNAw4QbW/7TC5vSZqBPY=",
-    "pQABAQoCoQBYIGs3vyxwmzEZ+qKNy4wpFkxc+Bgkb0D76ZEbxeeh/9DVA6EAoQBYILLf1TIwSB62q69Ojs/X1tzJ+dYHNAw4QbW/7TC5vSZqBPY=",
-    "pQABAQsCoQBYIGUiBeOJALP5dkMduUZ/u6MDhHNrsrBUa3f0YlSSWZbzA6EAoQBYILLf1TIwSB62q69Ojs/X1tzJ+dYHNAw4QbW/7TC5vSZqBPY=",
-    "pQABAQwCoQBYIMp6QNNTPDZgL3DSSD/QWWnBI7LsTZp2RhY/HLqnIwRZA6EAoQBYILLf1TIwSB62q69Ojs/X1tzJ+dYHNAw4QbW/7TC5vSZqBPY=",
-    "pQABAQ0CoQBYIJXSSUrE5RCNyB5pg+m6vGwK7RvJ+rs9dsdHitxnfDhuA6EAoQBYILLf1TIwSB62q69Ojs/X1tzJ+dYHNAw4QbW/7TC5vSZqBPY=",
-    "pQABAQ4CoQBYIHmtOX7jCKBHFDysb4H0z/QWoCSaEyjerZaT/HOP8bgDA6EAoQBYILLf1TIwSB62q69Ojs/X1tzJ+dYHNAw4QbW/7TC5vSZqBPY=",
-    "pQABAQ8CoQBYIIaMCTcPKj2HuYQ7i9ZaxUw9j5Bz8TPjoAaTZ5eB0w1kA6EAoQBYILLf1TIwSB62q69Ojs/X1tzJ+dYHNAw4QbW/7TC5vSZqBPY=",
-    "pQABARACoQBYIHWAOacKuWH81moJVveJ0FSfipWocfspOIBhaU6VLWUsA6EAoQBYILLf1TIwSB62q69Ojs/X1tzJ+dYHNAw4QbW/7TC5vSZqBPY=",
-    "pQABARECoQBYIA8XtUXtnMxQslULnNAeHBIivlLRe/+qdh2j6nTfDAchA6EAoQBYILLf1TIwSB62q69Ojs/X1tzJ+dYHNAw4QbW/7TC5vSZqBPY=",
-    "pQABARICoQBYIGgzg6SzgTTOgnk48pa6y2Rgjy004DkeBo4CMld3Jlr6A6EAoQBYILLf1TIwSB62q69Ojs/X1tzJ+dYHNAw4QbW/7TC5vSZqBPY=",
-    "pQABARMCoQBYIEoEFiIpCHgn74CAD+GhIfIgbQtdCqQqkOXHWxRlG6Y6A6EAoQBYILLf1TIwSB62q69Ojs/X1tzJ+dYHNAw4QbW/7TC5vSZqBPY=",
-    "pQABARQCoQBYINVEwTRxNSe0rxZxon4Rifz2l4rtQZn7mHtKYCiFAK9IA6EAoQBYILLf1TIwSB62q69Ojs/X1tzJ+dYHNAw4QbW/7TC5vSZqBPY=",
-    "pQABARUCoQBYIN3aeX2Ayi2rPFbiaYb+O2rdHUpFhzRs2j28pCmbGpflA6EAoQBYILLf1TIwSB62q69Ojs/X1tzJ+dYHNAw4QbW/7TC5vSZqBPY=",
-    "pQABARYCoQBYIJe5OJ17YKQrNmIH3sE++r++4Z5ld36axqAMjjQ3jtQWA6EAoQBYILLf1TIwSB62q69Ojs/X1tzJ+dYHNAw4QbW/7TC5vSZqBPY=",
-    "pQABARcCoQBYIASE94LjK6Raipk/lN/YewouqO+kcQGpxIqP+iW2hyHiA6EAoQBYILLf1TIwSB62q69Ojs/X1tzJ+dYHNAw4QbW/7TC5vSZqBPY=",
-    "pQABARgYAqEAWCBZ222LpS6/99Btlw+83PihrA655skwsNevt//8oz5axQOhAKEAWCCy39UyMEgetquvTo7P19bcyfnWBzQMOEG1v+0wub0magT2",
-    "pQABARgZAqEAWCDGEwo61w4O8T8lyw0HdoOjGWBKQUNqo6+jSfrPR9alrAOhAKEAWCCy39UyMEgetquvTo7P19bcyfnWBzQMOEG1v+0wub0magT2",
-    "pQABARgaAqEAWCBMSQoQ6B35plB80i1O3AWlJSftCEbCbju97Iykg5+NWQOhAKEAWCCy39UyMEgetquvTo7P19bcyfnWBzQMOEG1v+0wub0magT2"
-  ]
-
-someLastPrekeys :: [String]
-someLastPrekeys =
-  [ "pQABARn//wKhAFggnCcZIK1pbtlJf4wRQ44h4w7/sfSgj5oWXMQaUGYAJ/sDoQChAFgglacihnqg/YQJHkuHNFU7QD6Pb3KN4FnubaCF2EVOgRkE9g==",
-    "pQABARn//wKhAFggwO2any+CjiGP8XFYrY67zHPvLgp+ysY5k7vci57aaLwDoQChAFgglacihnqg/YQJHkuHNFU7QD6Pb3KN4FnubaCF2EVOgRkE9g==",
-    "pQABARn//wKhAFggoChErA5oTI5JT769hJV+VINmU8kougGdYqGd2U7hPa8DoQChAFgglacihnqg/YQJHkuHNFU7QD6Pb3KN4FnubaCF2EVOgRkE9g==",
-    "pQABARn//wKhAFggPLk4BBJ8THVLGm7r0K7EJITRlJnt6bpNzM9GTNRYcCcDoQChAFgglacihnqg/YQJHkuHNFU7QD6Pb3KN4FnubaCF2EVOgRkE9g==",
-    "pQABARn//wKhAFggqHASsRlZ1i8dESXRXBL2OvR+0yGUtqK9vJfzol1E+osDoQChAFgglacihnqg/YQJHkuHNFU7QD6Pb3KN4FnubaCF2EVOgRkE9g==",
-    "pQABARn//wKhAFggx/N1YhKXSJYJQxhWgHSA4ASaJKIHDJfmEnojfnp9VQ8DoQChAFgglacihnqg/YQJHkuHNFU7QD6Pb3KN4FnubaCF2EVOgRkE9g==",
-    "pQABARn//wKhAFggVL6QIpoqmtKxmB8HToiAPxfjSDEzJEUAoFKfhXou06YDoQChAFgglacihnqg/YQJHkuHNFU7QD6Pb3KN4FnubaCF2EVOgRkE9g==",
-    "pQABARn//wKhAFggRs74/ViOrHN+aS2RbGCwC0sJv1Sp/Q0pmRB15s9DCBMDoQChAFgglacihnqg/YQJHkuHNFU7QD6Pb3KN4FnubaCF2EVOgRkE9g==",
-    "pQABARn//wKhAFggtNO/hrwzt9M/1X6eK2sG6YFmA7BDqlFMEipbZOsg0vcDoQChAFgglacihnqg/YQJHkuHNFU7QD6Pb3KN4FnubaCF2EVOgRkE9g==",
-    "pQABARn//wKhAFgg1rZEY6vbAnEz+Ern5kRny/uKiIrXTb/usQxGnceV2HADoQChAFgglacihnqg/YQJHkuHNFU7QD6Pb3KN4FnubaCF2EVOgRkE9g==",
-    "pQABARn//wKhAFgg2647mOAVeOdhW57Q1zXDigDxRz/hB8ITFSZ7uo+pXH4DoQChAFgglacihnqg/YQJHkuHNFU7QD6Pb3KN4FnubaCF2EVOgRkE9g==",
-    "pQABARn//wKhAFggjddbHizABYOY0T6rvJeZCvV20dvTT9BYv95ri9bqSb8DoQChAFgglacihnqg/YQJHkuHNFU7QD6Pb3KN4FnubaCF2EVOgRkE9g==",
-    "pQABARn//wKhAFggCKT/GspZquUY6vKC4TFvaFqTH1QGG1ptauiaulnfqkUDoQChAFgglacihnqg/YQJHkuHNFU7QD6Pb3KN4FnubaCF2EVOgRkE9g==",
-    "pQABARn//wKhAFggv7bf/kEsTKFDGSgswsywq6AIxBq5AqZbLjDYDHfGjrcDoQChAFgglacihnqg/YQJHkuHNFU7QD6Pb3KN4FnubaCF2EVOgRkE9g==",
-    "pQABARn//wKhAFggUbjGhhh8EwZEPSz+Y31rYNUu7jsRR8dy1F5FSiJXfXEDoQChAFgglacihnqg/YQJHkuHNFU7QD6Pb3KN4FnubaCF2EVOgRkE9g==",
-    "pQABARn//wKhAFgg/4nz1uHiPBVGFvYjTMwGQ31bSFNctbU0r2nBtpsK9kcDoQChAFgglacihnqg/YQJHkuHNFU7QD6Pb3KN4FnubaCF2EVOgRkE9g==",
-    "pQABARn//wKhAFggwbJDyKl7T3+3Ihc0YF06Dz2J11My5qn7JKG+U+ti8lQDoQChAFgglacihnqg/YQJHkuHNFU7QD6Pb3KN4FnubaCF2EVOgRkE9g==",
-    "pQABARn//wKhAFgglc6nCoZR2/qjLp0tr7vRyuXqb7ugdHHDadjX7zSl4uMDoQChAFgglacihnqg/YQJHkuHNFU7QD6Pb3KN4FnubaCF2EVOgRkE9g==",
-    "pQABARn//wKhAFgg5ER8h0/bIADXjBXe/XPKdzekgv6nhJ4hp3vJ3jtTSbUDoQChAFgglacihnqg/YQJHkuHNFU7QD6Pb3KN4FnubaCF2EVOgRkE9g==",
-    "pQABARn//wKhAFggsgV6jq+GuNuvXk+ctHh570cNqEmfPhz34wcYCMCf9xIDoQChAFgglacihnqg/YQJHkuHNFU7QD6Pb3KN4FnubaCF2EVOgRkE9g==",
-    "pQABARn//wKhAFggdQdlPqkBw6+phKhohp3YaWQL710euZDnyMLFwf2cS0oDoQChAFgglacihnqg/YQJHkuHNFU7QD6Pb3KN4FnubaCF2EVOgRkE9g==",
-    "pQABARn//wKhAFggKlsI/snuQMoYcZRw/kN+BobPV5gwYeBClp0Wx9btTGUDoQChAFgglacihnqg/YQJHkuHNFU7QD6Pb3KN4FnubaCF2EVOgRkE9g==",
-    "pQABARn//wKhAFggtruFBClEgdPKvjpHsYLlWMev9L4OmYZwlxbY0NwvzOwDoQChAFgglacihnqg/YQJHkuHNFU7QD6Pb3KN4FnubaCF2EVOgRkE9g==",
-    "pQABARn//wKhAFggRUdh4cuYtFNL46RLnPy65goYInyreStKwsEcY3pPlLkDoQChAFgglacihnqg/YQJHkuHNFU7QD6Pb3KN4FnubaCF2EVOgRkE9g==",
-    "pQABARn//wKhAFggQtT7lLZzH171F4jCbHNwxEAt28FwdQ8Kt2tbxFzPgC0DoQChAFgglacihnqg/YQJHkuHNFU7QD6Pb3KN4FnubaCF2EVOgRkE9g==",
-    "pQABARn//wKhAFggQeUPM119c+6zRsEupA8zshTfrZiLpXx1Ji0UMMumq9IDoQChAFgglacihnqg/YQJHkuHNFU7QD6Pb3KN4FnubaCF2EVOgRkE9g=="
-  ]
 
 readServiceConfig :: Service -> App Value
 readServiceConfig srv = do
